test: consolidate 8 Plutus test files into PlutusData.test.ts
6 sections: Annotations, Compiler, Public API, Real-world types,
Edge cases, Benchmarks. 176 tests in one file, 273 total passing.
Removed duplicates and redundant enum comparison test.
